requirements.txt
boto3==1.34.59
email_validator==2.2.0
pandas==2.2.2
python-dotenv==1.0.1
